pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "mltemplate"
version = "0.0.1"
description = "mltemplate"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Allister Kohn", email = "contact@allisterkohn.com" }]
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
    "accelerate>=1.13.0",
    "awswrangler>=3.15.1",
    "boto3>=1.42.63",
    "catboost>=1.2.10",
    "dash>=4.0.0",
    "datasets>=4.6.1",
    "evaluate>=0.4.6",
    "fastapi>=0.135.1",
    "flask>=3.1.3",
    "gensim>=4.4.0",
    "imbalanced-learn>=0.14.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "joblib>=1.5.3",
    "jupyterlab>=4.5.5",
    "kaggle>=2.0.0",
    "keras>=3.13.2",
    "keras-tuner>=1.4.8",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "minio>=7.2.20",
    "mlflow>=3.10.1",
    "nltk>=3.9.3",
    "notebook>=7.5.4",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "polars>=1.38.1",
    "pre-commit>=4.5.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "spacy>=3.8.11",
    "tf-keras>=2.20.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchinfo>=1.8.0",
    "torchmetrics>=1.8.2",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["setup.py", "*.ipynb", ".venv", "__pycache__"]
cache-dir = "~/.cache/ruff"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = [
    "E501",
    "E731",
]

[tool.ruff.lint.isort]
known-first-party = ["mltemplate"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
cache_dir = "~/.cache/pytest"
